Adding a proper Browserify task (again) that combines files, and outputs to a temporary /out directory.
