+++ package.json
@@ -28,6 +28,7 @@
     "gulp-uncss": "^1.0.1",
     "gulp-watch": "~4.1.1",
     "run-sequence": "^1.0.2",
+    "vinyl-buffer": "^1.0.0",
     "vinyl-paths": "^1.0.0",
     "vinyl-source-stream": "^1.1.0",
     "vinyl-transform": "^1.0.0"
